First commited as after_9451_rev1

+++ src/main/java/net/spy/memcached/protocol/binary/ConcatenationOperationImpl.java
@@ -30,15 +30,15 @@ import net.spy.memcached.ops.OperationCallback;
 class ConcatenationOperationImpl extends SingleKeyOperationImpl implements
     ConcatenationOperation {
 
-  private static final int APPEND = 0x0e;
-  private static final int PREPEND = 0x0f;
+  private static final byte APPEND = 0x0e;
+  private static final byte PREPEND = 0x0f;
 
   private final long cas;
   private final ConcatenationType catType;
   private final byte[] data;
 
-  private static int cmdMap(ConcatenationType t) {
-    int rv;
+  private static byte cmdMap(ConcatenationType t) {
+    byte rv;
     switch (t) {
     case append:
       rv = APPEND;
