+++ src/main/java/net/spy/memcached/protocol/binary/FlushOperationImpl.java
@@ -28,14 +28,14 @@ import net.spy.memcached.ops.OperationCallback;
 
 class FlushOperationImpl extends OperationImpl implements FlushOperation {
 
-  private static final int CMD = 0x08;
+  private static final byte CMD = 0x08;
   private final int delay;
 
   public FlushOperationImpl(OperationCallback cb) {
-    this(0, cb);
+    this((byte)0, cb);
   }
 
-  public FlushOperationImpl(int d, OperationCallback cb) {
+  public FlushOperationImpl(byte d, OperationCallback cb) {
     super(CMD, generateOpaque(), cb);
     delay = d;
   }
